azure: generalize fsspec functionality into FSSpecWrapper
